NFS mount Cloud and SLES repos
In preparation for being able to serve them to managed nodes, NFS mount
the SLES and Cloud repositories from the mirror instead of configuring
them with HTTP. Since Ardana already serves repositories from
/opt/ardana_packager, symlink the mounts there after ardana-init sets it
up. We also need to ensure the nodes trust the Devel:Cloud:8 so that we
don't have to use disable_gpg_check in the ansible playbooks.

With mkcloud/crowbar, we mounted these repositories in /srv/tftpboot.
With Ardana, we must choose a different directory since the
deployerincloud* models install the dnsmasq package (for neutron) on the
deployer node, which forces /srv/tftpboot to be globally unreadable and
unsearchable.

For now, we still need to keep the HTTP-based repositories installed.
Once changes in osconfig-ansible[1] are merged, Ardana will be capable
of consuming these required repositories from the deployer node.

[1] https://gerrit.suse.provo.cloud/3061