Add github_pr configuration for Ardana test jobs
This commit adds configuration settings to trigger Jenkins jobs for
pull requests made to Ardana repositories forked to github.com/cmurphy.
This is for a proof of concept. It should not be used. The real thing
will most likely need to subscribe to Gerrit changes and not to GitHub
pull requests.

In this proof of concept, this configuration is used by this trigger
job[1] which is not committed to git.

[1] https://ci.suse.de/job/colleen-test-jenkins-gh-handlers