Add Jenkinsfile for Jenkins Pipeline POC
This commit adds a Jenkinsfile that supports building packages, running
the std-min integration test, reporting results, and submitting changes
for proposed code changes, proposed package changes, and already merged
changes. This is just a proof of concept, it should not be used as-is.

This Jenkinsfile is supporting this Jenkins Pipeline job[1] which is not
committed to git. This Pipeline job can be triggered by code changes
(currently only GitHub changes)[2], by IBS submitrequests[3], or
periodically/manually on already merged changes, in which case a
successful result would submit updates to the packages.

To do:
  * Stop hardcoding to the home:comurphy:* IBS namespace and the cmurphy
    GitHub namespace
  * Rewrite shell logic in python
  * Translate all Jenkins jobs to JJB syntax and commit them in git
  * Experiment with a real staging workflow rather than a fake staging
    project

[1] https://ci.suse.de/job/colleen-testpipeline
[2] https://ci.suse.de/job/colleen-test-jenkins-gh-handlers
[3] https://ci.suse.de/job/colleen-test-jenkins-obs-handlers